simplify organize output
Removes methods document and script README.
Renames CwlReport to ReadmeReport.
Simplifies top level README.
